\name{biasVardistPsi0Gradient}
\Rdversion{1.0}
\alias{biasVardistPsi0Gradient}
\title{one line description}
\description{
  description here.
}
\usage{
values <- biasVardistPsi0Gradient(biaskern, vardist, Z, covGrad)
}
\arguments{
  \item{biaskern}{the kernel structure associated with the bias kernel.}
  \item{vardist}{}
  \item{Z}{}
  \item{covGrad}{}
}
\value{
  \item{gKern}{}
  \item{gVarmeans}{}
  \item{gVarcovars}{}
  \item{gInd}{}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
