\name{biasVardistPsi1Compute}
\Rdversion{1.0}
\alias{biasVardistPsi1Compute}
\title{one line description}
\description{
  description
}
\usage{
values <- biasVardistPsi1Compute(biasKern, vardist, Z)
}
\arguments{
  \item{biasKern}{the kernel structure associated with the white kernel.}
  \item{vardist}{description}
  \item{Z}{description}
}
\value{
  \item{K}{description}
  \item{P}{description}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
