\name{biasVardistPsi2Compute}
\Rdversion{1.0}
\alias{biasVardistPsi2Compute}
\title{one line description}
\description{
  description
}
\usage{
values <- biasVardistPsi2Compute(biasKern, vardist, Z)
}
\arguments{
  \item{biasKern}{the kernel structure associated with the bias kernel.}
  \item{vardist}{description}
  \item{Z}{description}
}
\value{
  \item{Psi2}{description}
  \item{P}{description}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
