\name{biasVardistPsi2Gradient}
\Rdversion{1.0}
\alias{biasVardistPsi2Gradient}
\title{Compute gradient of bias variational PSI2.}
\description{
  description here.
}
\usage{
values <- biasVardistPsi2Gradient(biaskern, vardist, Z, covGrad)
}
\arguments{
  \item{biaskern}{the kernel structure associated with the bias kernel.}
  \item{vardist}{}
  \item{Z}{}
  \item{covGrad}{}
}
\value{
  \item{gKern}{}
  \item{gVarmeans}{}
  \item{gVarcovars}{}
  \item{gInd}{}
}
\seealso{
\code{\link{}}.
}
\keyword{model}
