\name{dist2}
\Rdversion{1.0}
\alias{dist2}
\title{Calculates squared distance between two sets of points.}
\description{
   D = DIST2(X, C) takes two matrices of vectors and calculates the squared Euclidean distance between them.  Both matrices must be of the same column dimension.  If X has M rows and N columns, and C has L rows and N columns, then the result has M rows and L columns.  The I, Jth entry is the  squared distance from the Ith row of X to the Jth row of C.
}
\usage{
dist2()
}
\arguments{
}
\value{
}
\keyword{model}
