\name{eigdec}
\Rdversion{1.0}
\alias{eigdec}
\title{ Sorted eigendecomposition}
\description{
   EVALS = EIGDEC(X, N computes the largest N eigenvalues of the matrix X in descending order.  [EVALS, EVEC] = EIGDEC(X, N) also computes the corresponding eigenvectors.
}
\usage{
eigdec()
}
\arguments{
}
\value{
}
\keyword{model}
