\name{getSymbols}
\Rdversion{1.0}
\alias{getSymbols}
\title{Get a cell array of different plot symbols.}
\description{
  returns a cell array of different plot symbols. A maximum of 66 distinct symbols will be created.
}
\usage{
symbol <- getSymbols(number)
}
\arguments{
  \item{number}{the number of different plot symbols required.}
}
\value{
  \item{symbol}{cell array of the different symbols.}
}
\seealso{
\code{\link{plot}}.
}
\keyword{model}
