\name{linard2biasVardistPsi2Compute}
\Rdversion{1.0}
\alias{linard2biasVardistPsi2Compute}
\title{one line description}
\description{
  description
}
\usage{
values <- linard2biasVardistPsi2Compute(linard2Kern, biasKern,
                                        vardist, Z)
}
\arguments{
  \item{linard2Kern}{the kernel structure associated with the linard2 kernel.}
  \item{biasKern}{description}
  \item{vardist}{description}
  \item{Z}{description}
}
\value{
  \item{Psi2}{description}
  \item{Pnobias}{description}
  \item{Psi1}{description}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
