\name{lvmPrintPlot}
\Rdversion{1.0}
\alias{lvmPrintPlot}
\title{Print latent space for learnt model.}
\description{
  prints a latent space repsresentation for an LVM model.
}
\usage{
lvmPrintPlot(model, lbls, capName, experimentNo)
}
\arguments{
  \item{model}{the model to use for plotting the latent space.}
  \item{lbls}{any lables that are available for plotting.}
  \item{capName}{the name of the saved plots.}
  \item{experimentNo}{the experiment number to assign to the files. }
}
\value{
}
\seealso{
\code{\link{lvmScatterPlot}}.
}
\keyword{model}
