\name{lvmTwoDPlot}
\Rdversion{1.0}
\alias{lvmTwoDPlot}
\title{Helper function for plotting the labels in 2-D.}
\description{
  helper function for plotting an embedding in 2-D with symbols.
}
\usage{
lvmTwoDPlot(X, lbl, symbol)
}
\arguments{
  \item{X}{the data to plot.}
  \item{lbl}{the labels of the data point.}
  \item{symbol}{the symbols to use for the different labels.}
}
\value{
}
\seealso{
\code{\link{lvmScatterPlot}, \link{lvmVisualise}}.
}
\keyword{model}
