\name{printPlot}
\Rdversion{1.0}
\alias{printPlot}
\title{Print a plot to eps and png files.}
\description{
  prints a plot to the specified file name and directory.
}
\usage{
printPlot(fileName, directory, directoryPng)
}
\arguments{
  \item{fileName}{the base name of the file to use.}
  \item{directory}{the directory to place the eps files in.}
  \item{directoryPng}{the directory to place png the file in. }
}
\value{
}
\seealso{
\code{\link{preparePlot}}.
}
\keyword{model}
