\name{rbfard2VardistPsi2Compute}
\Rdversion{1.0}
\alias{rbfard2VardistPsi2Compute}
\title{one line description}
\description{
  description
}
\usage{
values <- rbfard2VardistPsi2Compute(rbfardKern, vardist, Z)
}
\arguments{
  \item{rbfardKern}{the kernel structure associated with the rbfard2 kernel.}
  \item{vardist}{description}
  \item{Z}{description}
}
\value{
  \item{K}{description}
  \item{outKern}{description}
  \item{sumKern}{description}
  \item{Kgvar}{description}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
