\name{vardistCreate}
\Rdversion{1.0}
\alias{vardistCreate}
\title{creates the structure of the variational distirbution over the latent values in the GP-LVM}
\description{
   The variational distribution is assumed to be factorized over the the latent variables. Each factor each a Gaussain N(x_n|mu_n,S_n) with a diagonal covariance matrix
}
\usage{
vardistCreate()
}
\arguments{
}
\value{
}
\keyword{model}
