\name{vargplvmUpdateStats}
\Rdversion{1.0}
\alias{vargplvmUpdateStats}
\title{Update stats for VARGPLVM model.}
\description{
  no description available.
}
\usage{
vargplvmUpdateStats()
}
\arguments{
}
\value{
}
\seealso{
\code{\link{vargplvmOptimise}, \link{vargplvmExpandParam}}.
}
\keyword{model}
