\name{whiteVardistPsi0Gradient}
\Rdversion{1.0}
\alias{whiteVardistPsi0Gradient}
\title{one line description}
\description{
  description here.
}
\usage{
values <- whiteVardistPsi0Gradient(whitekern, vardist, Z,
                                   covGrad)
}
\arguments{
  \item{whitekern}{the kernel structure associated with the white kernel.}
  \item{vardist}{}
  \item{Z}{}
  \item{covGrad}{}
}
\value{
  \item{gKern}{}
  \item{gVarmeans}{}
  \item{gVarcovars}{}
  \item{gInd}{}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
