\name{whiteVardistPsi1Gradient}
\Rdversion{1.0}
\alias{whiteVardistPsi1Gradient}
\title{Compute gradient of white variational Psi1.}
\description{
  description here.
}
\usage{
values <- whiteVardistPsi1Gradient(whitekern, vardist, Z,
                                   covGrad)
}
\arguments{
  \item{whitekern}{the kernel structure associated with the white kernel.}
  \item{vardist}{}
  \item{Z}{}
  \item{covGrad}{}
}
\value{
  \item{gKern}{}
  \item{gVarmeans}{}
  \item{gVarcovars}{}
  \item{gInd}{}
}
\seealso{
\code{\link{}}.
}
\keyword{model}
